Fix error when searching for "on"
Perhaps the "locations" thing lists all the places something is
exported, including re-exports.  The old code would throw an error
unless there was exactly one thing in the "locations" list. For now, it
seems to make the most sense to treat each one as a separate result
instead.

The actual data causing the error was one of the results for "on", which
had two values in the locations list: AssocNone from both string-parsers
and parsing.