Switch to stack
Mainly because I'm fed up with my sandbox getting confused and having to
blow it all away and rebuild it all so often, and because 'cabal freeze'
is painful.

This also upgrades a bunch of packages, as it's now based on lts-5.3.
Most of these upgrades are minor or patch-level (that is, components 3
or 4), so we probably don't need to worry about them. It also means
we're now using GHC 7.10.3 instead of 7.8.4.