Clearly scoped .message component
Start to use BEM naming pattern for modules. This pattern is controversial among some groups, however, at its core it's just a human-readable method of name mangling and that's what we're using it for. Due to the everything-is-global nature of CSS we need some form of scoping to prevent rules from leaking from one to another component.

Also removed .success modifier and colors because it is not used anywhere.