Clarify that packages are published to Pursuit
I introduced the term "publish" instead of upload for clarity (Pursuit is about publishing documentation for a package). This doesn't distinguish between the person who merely uploaded a package to Pursuit (could be anyone) and the person who published the package (made the Git tag). So I wanted to clarify that you "publish to Pursuit", keep the term "Publish" to describe this action, but still keep "uploaded by" to label the person who put the documentation on Pursuit.