Add favicon
This commit uses a 32x32 PNG favicon instead of an ICO because the ICO turned out to be much bigger (15kB for the ICO vs 350B for the PNG, of course the ICO includes sizes 16/32/48) due to the alpha transparency we want to use.

Alternatively, we could just serve multiple PNGs like we do on the Purescript website. Here I failed in figuring out how to do this with Yesod :(

Preferrably, we would also include touch icons and the like to cater for all platforms. However, here too I do not know how to do that with Yesod (we'd have to have routes to SVG, JSON, XML, PNG, and ICO).