Use static xml/json files instead of generated ones
The benefit of generating these files is that the compiler can assure us that the files we point to are actually available. Depending on the configuration, the server can also place them whereever it wants because it controls the URLs.

However, this introduces a lot of unnecessary machinery, so we favored doing it with hard-coded paths instead. This way we can also stay closer to what is in the purescript-logo repository.

Also removed documentation from the static folder, as that would get published. Added a short note in README instead.