'Load more' button for search results (fixes #305) (#344)
If there are more search results than fit in one 'page' (= 50 results),
Pursuit now displays a 'load more' link at the end of the results;
clicking this loads another page of results via AJAX and appends them to
the document. If JavaScript is disabled, the 'load more' link instead
loads the same page, except with another 50 results (or however many
remaining results there are, if less than 50).

We distinguish the cases where there are no more results vs where
Pursuit has decided not to show any more results (for the sake of
minimising load on the server) by displaying an explanatory message at
the end of the search results.

The JSON output only contains up to 50 results, so subsequent pages of
results must be fetched with a new request.

We also include a Link header in the response, referencing the next and
previous URLs, mainly for clients using JSON (although clients using
HTML might find it useful too).