Structure help menu with existing styles
We currently don't have styles for a <ul> nested within a <dl>, so the spacing is off and it doesn't look right. Instead of introducing a new custom list style for the help index, I try to use the existing grouped_list for this.

The downside of this approach is that the help menu doesn't look like one single nested list, but like several lists. For the content that is there at the moment this works well, though, in my opinion.