Use proper typographic quotes
We might not want to keep this commit because people might not know how to write proper typographic quotes and a mix of straight and real quotes would be worse.

However, typographic quotes do look better and also say “professional” in a subtle way that most people will not be able to put their finger on but still notice.

An alternative approach could be to enable a smart quote filter for the Markdown documents.